Consolidate memory client modules
