+++ tests/test_quantum_memory_server.py
@@ -3,7 +3,7 @@ import numpy as np
 
 from asi.vector_store import VectorStore
 from asi.quantum_memory_server import QuantumMemoryServer
-from asi.quantum_memory_client import QuantumMemoryClient
+from asi.memory_clients import QuantumMemoryClient
 
 
 class TestQuantumMemoryServer(unittest.TestCase):
